moving summary into the Features folder
