+++ webui/app-tracker/src/components/Menu/Menu.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import "./menu.css";
+import "./Menu.css";
 
 // hard coded without functionality while I short out the layout.
 export default function Menu() {
